Allow transmit length greater than 255

+++ inc/lib_ke_protocol.h
@@ -42,7 +42,7 @@
 #define KE_TIMEOUT                   5000
 #define MAX_RETRIES                  25
 
-typedef int (*TRANSMIT_KE_DATA)(uint8_t *data, uint8_t len);
+typedef int (*TRANSMIT_KE_DATA)(uint8_t *data, uint32_t len);
 typedef void (*REQUEST_ACTIVE_COOLING)(uint8_t level);
 typedef uint32_t (*KE_STRING_TO_JSON)(char *buffer, uint32_t buffer_size);
 typedef bool (*KE_JSON_TO_FUNCTION)(const char *json_str);
